requirements.txt
aiogram==3.19.0
asyncpg==0.30.0
playwright==1.51.0
python-dotenv==1.1.0
